requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==4.1
django-coverage-plugin==1.3.1
django-extensions==1.6.7
pylint==1.6.4
pylint-common==0.2.2
pylint-django==0.7.2
pylint-celery==0.3
pylint-plugin-utils==0.2.4
Werkzeug==0.11.10
django-test-plus==1.0.14
factory_boy==2.7.0
django-debug-toolbar==1.5

# improved REPL
ipdb==0.10.1

pytest-django==2.9.1
pytest-sugar==0.7.1
moto==0.4.30
responses==0.5.1
